also add docs for hovered events
