--- crates/elements/src/_docs/events/globalkeydown.md
@@ -1,16 +1,0 @@
-The `globalkeydown` event fires when the user starts pressing any key.
-
-Event Data: [`KeyboardData`](crate::events::KeyboardData)
-
-### Example
-
-```rust, no_run
-# use freya::prelude::*;
-fn app() -> Element {
-    rsx!(
-        rect {
-            onglobalkeydown: |e| println!("Event: {e:?}")
-        }
-    )
-}
-```
