--- crates/elements/src/_docs/events/keydown.md
@@ -1,16 +1,0 @@
-The `keydown` event fires when the user starts pressing any key in the currently focused element.
-
-Event Data: [`KeyboardData`](crate::events::KeyboardData)
-
-### Example
-
-```rust, no_run
-# use freya::prelude::*;
-fn app() -> Element {
-    rsx!(
-        rect {
-            onkeydown: |e| println!("Event: {e:?}")
-        }
-    )
-}
-```
