--- crates/elements/src/_docs/events/touchmove.md
@@ -1,19 +1,0 @@
-The `touchmove` event fires when the user is touching over an element.
-
-Event Data: [`TouchData`](crate::events::TouchData)
-
-### Example
-
-```rust, no_run
-# use freya::prelude::*;
-fn app() -> Element {
-    rsx!(
-        rect {
-            width: "100",
-            height: "100",
-            background: "red",
-            ontouchmove: |_| println!("Touching!")
-        }
-    )
-}
-```
