+++ crates/elements/src/elements.rs
@@ -1,93 +1,4 @@
-pub use events::*;
-
-macro_rules! def_element {
-    (
-        $(
-            $(#[$attr:meta])*
-            $name:ident {
-                $(
-                    $(#[$attr_method:meta])*
-                    $fil:ident,
-                )*
-            };
-         )*
-        ) => {
-        $(
-            impl_element!(
-                $(#[$attr])*
-                $name {
-                    $(
-                        $(#[$attr_method])*
-                        $fil,
-                    )*
-                };
-            );
-        )*
-
-        /// This module contains helpers for rust analyzer autocompletion
-        #[doc(hidden)]
-        pub mod completions {
-            /// This helper tells rust analyzer that it should autocomplete the element name with braces.
-            #[allow(non_camel_case_types)]
-            pub enum CompleteWithBraces {
-                $(
-                    $(#[$attr])*
-                    $name {}
-                ),*
-            }
-        }
-    };
-}
-
-macro_rules! impl_element {
-    (
-        $(
-            $(#[$attr:meta])*
-            $name:ident {
-                $(
-                    $(#[$attr_method:meta])*
-                    $fil:ident,
-                )*
-            };
-         )*
-    ) => {
-        $(
-            #[allow(non_camel_case_types)]
-            $(#[$attr])*
-            pub mod $name {
-                #[doc(hidden)]
-                pub const TAG_NAME: &'static str = stringify!($name);
-                #[doc(hidden)]
-                pub const NAME_SPACE: Option<&'static str> = None;
-
-                $(
-                   pub use crate::attributes::$fil::$fil;
-                )*
-            }
-        )*
-    };
-}
-
-#[macro_export]
-macro_rules! def_attribute {
-    (
-        $(
-            $(#[$attr:meta])*
-            $fil:ident,
-         )*
-    ) => {
-        $(
-            #[allow(non_camel_case_types)]
-
-            pub mod $fil {
-
-                #[allow(non_upper_case_globals)]
-                $(#[$attr])*
-                pub const $fil: (&'static str, Option<&'static str>, bool) = (stringify!($fil), None, false);
-            }
-        )*
-    };
-}
+use crate::def_element;
 
 def_element!(
     /// `rect` is a generic element that acts as a container for other elements.
@@ -767,167 +678,3 @@ def_element!(
         a11y_vertical_offset,
     };
 );
-
-pub mod events {
-    use std::any::Any;
-
-    use crate::events::*;
-
-    #[doc(hidden)]
-    pub trait EventReturn<P>: Sized {
-        fn spawn(self) {}
-    }
-
-    impl EventReturn<()> for () {}
-    #[doc(hidden)]
-    pub struct AsyncMarker;
-
-    impl<T> EventReturn<AsyncMarker> for T
-    where
-        T: std::future::Future<Output = ()> + 'static,
-    {
-        #[inline]
-        fn spawn(self) {
-            dioxus_core::prelude::spawn(self);
-        }
-    }
-
-    /// A platform specific event.
-    #[doc(hidden)]
-    pub struct ErasedEventData {
-        event: Box<dyn Any>,
-    }
-
-    impl ErasedEventData {
-        pub fn new(event: Box<dyn Any>) -> Self {
-            Self { event }
-        }
-
-        pub fn downcast<T: 'static>(&self) -> Option<&T> {
-            self.event.downcast_ref::<T>()
-        }
-
-        pub fn downcast_mut<T: 'static>(&mut self) -> Option<&mut T> {
-            self.event.downcast_mut::<T>()
-        }
-
-        pub fn into_inner<T: 'static>(self) -> Option<T> {
-            self.event.downcast::<T>().ok().map(|e| *e)
-        }
-    }
-
-    macro_rules! impl_event {
-        (
-            $data:ty;
-            $(
-                $( #[$attr:meta] )*
-                $name:ident $(: $event:literal)?
-            )*
-        ) => {
-            $(
-                $( #[$attr] )*
-                #[inline]
-                pub fn $name<__Marker>(mut _f: impl ::dioxus_core::prelude::SuperInto<::dioxus_core::prelude::EventHandler<::dioxus_core::Event<$data>>, __Marker>) -> ::dioxus_core::Attribute {
-                    // super into will make a closure that is owned by the current owner (either the child component or the parent component).
-                    // We can't change that behavior in a minor version because it would cause issues with Components that accept event handlers.
-                    // Instead we run super into with an owner that is moved into the listener closure so it will be dropped when the closure is dropped.
-                    let owner = <::generational_box::UnsyncStorage as ::generational_box::AnyStorage>::owner();
-                    let event_handler = ::dioxus_core::prelude::with_owner(owner.clone(), || _f.super_into());
-                    ::dioxus_core::Attribute::new(
-                        impl_event!(@name $name $($event)?),
-                        ::dioxus_core::AttributeValue::listener(move |e: ::dioxus_core::Event<crate::ErasedEventData>| {
-                            // Force the owner to be moved into the event handler
-                            _ = &owner;
-                            event_handler.call(e.map(|e| e.into()));
-                        }),
-                        None,
-                        false,
-                    ).into()
-                }
-
-                #[doc(hidden)]
-                $( #[$attr] )*
-                pub mod $name {
-                    use super::*;
-
-                    // When expanding the macro, we use this version of the function if we see an inline closure to give better type inference
-                    $( #[$attr] )*
-                    pub fn call_with_explicit_closure<
-                        __Marker,
-                        Return: ::dioxus_core::SpawnIfAsync<__Marker> + 'static,
-                    >(
-                        event_handler: impl FnMut(::dioxus_core::Event<$data>) -> Return + 'static,
-                    ) -> ::dioxus_core::Attribute {
-                        #[allow(deprecated)]
-                        super::$name(event_handler)
-                    }
-                }
-            )*
-        };
-
-        (@name $name:ident) => {
-            stringify!($name)
-        };
-    }
-
-    impl_event! [
-        MouseData;
-
-        onclick
-        onglobalclick
-        onmiddleclick
-        onrightclick
-        onmouseup
-        onmousedown
-        onglobalmousedown
-        onmousemove
-        onglobalmousemove
-        onmouseleave
-        onmouseenter
-    ];
-
-    impl_event! [
-        WheelData;
-
-        onwheel
-    ];
-
-    impl_event! [
-        KeyboardData;
-
-        onkeydown
-
-        onkeyup
-
-        onglobalkeydown
-        onglobalkeyup
-    ];
-
-    impl_event! [
-        TouchData;
-
-        ontouchcancel
-        ontouchend
-        ontouchmove
-        ontouchstart
-    ];
-
-    impl_event! [
-        PointerData;
-
-        onpointerdown
-        onpointerup
-        onglobalpointerup
-        onpointermove
-        onpointerenter
-        onpointerleave
-    ];
-
-    impl_event! [
-        FileData;
-
-        onfiledrop
-        onglobalfilehover
-        onglobalfilehovercancelled
-    ];
-}
